move VNode.ts
