+++ packages/core/src/VNodes/RootNode.ts
@@ -1,4 +1,4 @@
-import { VNode, VNodeType } from '../VNode';
+import { VNode, VNodeType } from './VNode';
 
 export class RootNode extends VNode {
     constructor() {
