+++ packages/plugin-devtools/src/components/InfoComponent.ts
@@ -1,6 +1,6 @@
 import { OwlUIComponent } from '../../../owl-ui/src/OwlUIComponent';
 import { VRange } from '../../../core/src/VRange';
-import { VNode } from '../../../core/src/VNode';
+import { VNode } from '../../../core/src/VNodes/VNode';
 
 interface InfoState {
     currentTab: string;
